DCS-248 split out ro and ca/dm notification sending
