+++ server/services/notifications/caAndDmNotificationSender.js
@@ -1,18 +1,14 @@
-const moment = require('moment')
 const R = require('ramda')
-const templates = require('./config/notificationTemplates')
-const logger = require('../../log.js')
-const { getIn, isEmpty } = require('../utils/functionalHelpers')
-const { getRoCaseDueDate, getRoNewCaseDueDate } = require('../utils/dueDates')
+const logger = require('../../../log.js')
+const { getIn, isEmpty } = require('../../utils/functionalHelpers')
 
 module.exports = function createNotificationService(
   prisonerService,
   roContactDetailsService,
   configClient,
-  notifyClient,
-  audit,
+  notificationSender,
   nomisClientBuilder,
-  { notifications: { notifyKey, activeNotificationTypes, clearingOfficeEmail, clearingOfficeEmailEnabled }, domain }
+  { notifications: { activeNotificationTypes, clearingOfficeEmail, clearingOfficeEmailEnabled }, domain }
 ) {
   async function getNotificationData({
     prisoner,
@@ -85,11 +81,6 @@ module.exports = function createNotificationService(
     ]
   }
 
-  const getRoAndClearingOfficeNotifications = async args => {
-    const notifications = await getRoNotifications(args)
-    return [...notifications, ...clearingOfficeNotification(notifications)]
-  }
-
   async function getCaNotifications({ common, token, submissionTarget, sendingUserName }) {
     const agencyId = getIn(submissionTarget, ['agencyId'])
 
@@ -116,40 +107,6 @@ module.exports = function createNotificationService(
         templateName: 'STANDARD',
       }
     })
-  }
-
-  async function getRoNotifications({ common, token, submissionTarget, bookingId, transitionDate }) {
-    const deliusId = getIn(submissionTarget, ['deliusId'])
-    if (isEmpty(deliusId)) {
-      logger.error('Missing COM deliusId')
-      return []
-    }
-
-    const [establishment, roContactInfo] = await Promise.all([
-      prisonerService.getEstablishmentForPrisoner(bookingId, token),
-      roContactDetailsService.getContactDetails(deliusId),
-    ])
-
-    if (!roContactInfo) {
-      logger.error(`No contact info for delius id: ${deliusId}`)
-      return []
-    }
-
-    const prison = getIn(establishment, ['premise'])
-
-    if (isEmpty(prison)) {
-      logger.error(`Missing prison for bookingId: ${bookingId}`)
-      return []
-    }
-
-    const date = transitionDate ? getRoCaseDueDate(moment(transitionDate)) : getRoNewCaseDueDate()
-    const { email, orgEmail, organisation } = roContactInfo
-    const base = { personalisation: { ...common, ro_name: submissionTarget.name, organisation, prison, date } }
-
-    return [
-      ...(isEmpty(email) ? [] : [{ ...base, email, templateName: 'STANDARD' }]),
-      ...(isEmpty(orgEmail) ? [] : [{ ...base, email: orgEmail, templateName: 'COPY' }]),
-    ]
   }
 
   async function getDmNotifications({ common, token, bookingId }) {
@@ -179,22 +136,6 @@ module.exports = function createNotificationService(
     CA_DECISION: R.compose(
       R.then(replaceTemplateNames({ STANDARD: 'CA_DECISION' })),
       getCaNotifications
-    ),
-    RO_NEW: R.compose(
-      R.then(replaceTemplateNames({ STANDARD: 'RO_NEW', COPY: 'RO_NEW_COPY' })),
-      getRoAndClearingOfficeNotifications
-    ),
-    RO_TWO_DAYS: R.compose(
-      R.then(replaceTemplateNames({ STANDARD: 'RO_TWO_DAYS', COPY: 'RO_TWO_DAYS_COPY' })),
-      getRoNotifications
-    ),
-    RO_DUE: R.compose(
-      R.then(replaceTemplateNames({ STANDARD: 'RO_DUE', COPY: 'RO_DUE_COPY' })),
-      getRoNotifications
-    ),
-    RO_OVERDUE: R.compose(
-      R.then(replaceTemplateNames({ STANDARD: 'RO_OVERDUE', COPY: 'RO_OVERDUE_COPY' })),
-      getRoAndClearingOfficeNotifications
     ),
     DM_NEW: R.compose(
       R.then(replaceTemplateNames({ STANDARD: 'DM_NEW' })),
@@ -230,7 +171,7 @@ module.exports = function createNotificationService(
         return []
       }
 
-      await notify({
+      await notificationSender.notify({
         sendingUserName,
         notificationType,
         bookingId,
@@ -247,53 +188,8 @@ module.exports = function createNotificationService(
     }
   }
 
-  async function notify({ sendingUserName, notificationType, bookingId, notifications } = {}) {
-    if (isEmpty(notifyKey) || notifyKey === 'NOTIFY_OFF') {
-      logger.warn('No notification API key - notifications disabled')
-      return
-    }
-
-    if (isEmpty(notifications)) {
-      logger.warn('Empty notifications')
-      return
-    }
-
-    notifications.forEach(notification => {
-      if (isEmpty(templates[notification.templateName])) {
-        logger.warn(`Unmapped notification template name: ${notification.templateName}`)
-        return
-      }
-
-      const { templateId } = templates[notification.templateName]
-
-      if (isEmpty(notification.email)) {
-        logger.warn('Empty notification email')
-      } else {
-        notifyClient
-          .sendEmail(templateId, notification.email, { personalisation: notification.personalisation })
-          .then(() => {
-            logger.info(`Successful notify for email: ${notification.email}`)
-          })
-          .catch(error => {
-            logger.error('Error sending notification email ', notification.email)
-            logger.error(error.stack)
-          })
-      }
-    })
-    auditEvent(sendingUserName, bookingId, notificationType, notifications)
-  }
-
-  function auditEvent(user, bookingId, notificationType, notifications) {
-    audit.record('NOTIFY', user, {
-      bookingId,
-      notificationType,
-      notifications,
-    })
-  }
-
   return {
     getNotificationData,
     sendNotifications,
-    notify,
   }
 }
