+++ test/routes/sendTest.js
@@ -2,31 +2,25 @@ const request = require('supertest')
 const {
   createLicenceServiceStub,
   createPrisonerServiceStub,
-  createNotificationServiceStub,
-  createUserAdminServiceStub,
   appSetup,
   auditStub,
   createSignInServiceStub,
+  users,
 } = require('../supertestSetup')
 
 const standardRouter = require('../../server/routes/routeWorkers/standardRouter')
 const createRoute = require('../../server/routes/send')
+const transitionsForDestinations = require('../../server/services/notifications/transitionsForDestinations')
 
 describe('send', () => {
-  let prisonerService
-  let licenceService
-  let userAdminService
   let notificationService
+  let prisonerService
 
   const prisoner = { firstName: 'first', lastName: 'last', dateOfBirth: 'off-dob', offenderNo: 'AB1234A' }
   const submissionTarget = { premise: 'HMP Blah', agencyId: 'LT1', name: 'Something', deliusId: 'delius' }
 
   beforeEach(() => {
-    licenceService = createLicenceServiceStub()
     prisonerService = createPrisonerServiceStub()
-    userAdminService = createUserAdminServiceStub()
-    notificationService = createNotificationServiceStub()
-
     prisonerService.getOrganisationContactDetails = sinon.stub().resolves(submissionTarget)
 
     prisonerService.getEstablishmentForPrisoner = sinon
@@ -35,17 +29,14 @@ describe('send', () => {
 
     prisonerService.getPrisonerPersonalDetails = sinon.stub().resolves(prisoner)
 
-    userAdminService.getRoUserByDeliusId = sinon.stub().resolves({ orgEmail: 'expected@email' })
-
-    notificationService.sendNotifications = sinon.stub().resolves({})
-
-    auditStub.record.reset()
-    notificationService.sendNotifications.reset()
+    notificationService = {
+      send: sinon.stub().resolves({}),
+    }
   })
 
   describe('Get send/:destination/:bookingId', () => {
     it('renders caToRo form when addressReview is destination', () => {
-      const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'caUser')
+      const app = createApp({ prisonerServiceStub: prisonerService }, 'caUser')
       return request(app)
         .get('/hdc/send/addressReview/123')
         .expect(200)
@@ -56,7 +47,7 @@ describe('send', () => {
     })
 
     it('renders caToRo form when bassReview is destination', () => {
-      const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'caUser')
+      const app = createApp({ prisonerServiceStub: prisonerService }, 'caUser')
       return request(app)
         .get('/hdc/send/bassReview/123')
         .expect(200)
@@ -67,7 +58,7 @@ describe('send', () => {
     })
 
     it('renders roToCa form when finalChecks is destination', () => {
-      const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'roUser')
+      const app = createApp({ prisonerServiceStub: prisonerService }, 'roUser')
       return request(app)
         .get('/hdc/send/finalChecks/123')
         .expect(200)
@@ -78,7 +69,7 @@ describe('send', () => {
     })
 
     it('renders caToDm form when approval is destination', () => {
-      const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'caUser')
+      const app = createApp({ prisonerServiceStub: prisonerService }, 'caUser')
       return request(app)
         .get('/hdc/send/approval/123')
         .expect(200)
@@ -89,7 +80,7 @@ describe('send', () => {
     })
 
     it('renders dmToCa form when decided is destination', () => {
-      const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'dmUser')
+      const app = createApp({ prisonerServiceStub: prisonerService }, 'dmUser')
       return request(app)
         .get('/hdc/send/decided/123')
         .expect(200)
@@ -100,7 +91,7 @@ describe('send', () => {
     })
 
     it('renders caToDmRefusal form when refusal is destination', () => {
-      const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'caUser')
+      const app = createApp({ prisonerServiceStub: prisonerService }, 'caUser')
       return request(app)
         .get('/hdc/send/refusal/123')
         .expect(200)
@@ -111,7 +102,7 @@ describe('send', () => {
     })
 
     it('renders dmToCaReturn form when return is destination', () => {
-      const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'dmUser')
+      const app = createApp({ prisonerServiceStub: prisonerService }, 'dmUser')
       return request(app)
         .get('/hdc/send/return/123')
         .expect(200)
@@ -122,7 +113,7 @@ describe('send', () => {
     })
 
     it('gets a submission target for caToRo', () => {
-      const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'caUser')
+      const app = createApp({ prisonerServiceStub: prisonerService }, 'caUser')
       return request(app)
         .get('/hdc/send/addressReview/123')
         .expect(200)
@@ -133,7 +124,7 @@ describe('send', () => {
     })
 
     it('gets a submission target for roToCa', () => {
-      const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'roUser')
+      const app = createApp({ prisonerServiceStub: prisonerService }, 'roUser')
       return request(app)
         .get('/hdc/send/finalChecks/123')
         .expect(200)
@@ -144,7 +135,7 @@ describe('send', () => {
     })
 
     it('should throw if get requested by wrong user', () => {
-      const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'roUser')
+      const app = createApp({ prisonerServiceStub: prisonerService }, 'roUser')
       return request(app)
         .get('/hdc/send/refusal/123')
         .expect(403)
@@ -153,135 +144,46 @@ describe('send', () => {
 
   describe('POST send/:destination/:bookingId', () => {
     describe('Sending', () => {
-      it('calls markForHandover via licenceService for addressReview', () => {
-        const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService })
+      it('ca user sends to ro to review address ', () => {
+        const app = createApp({ prisonerServiceStub: prisonerService, notificationServiceStub: notificationService })
 
         return request(app)
           .post('/hdc/send/addressReview/123')
           .expect(() => {
-            expect(prisonerService.getOrganisationContactDetails).to.be.calledOnce()
-            expect(prisonerService.getOrganisationContactDetails).to.be.calledWith('RO', '123', 'token')
-            expect(licenceService.markForHandover).to.be.calledOnce()
-            expect(licenceService.markForHandover).to.be.calledWith('123', 'caToRo')
-          })
-      })
-
-      it('uses system token form RO', () => {
-        const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'roUser')
-
-        return request(app)
-          .post('/hdc/send/finalChecks/123')
-          .expect(() => {
-            expect(prisonerService.getOrganisationContactDetails).to.be.calledWith('CA', '123', 'system-token')
-          })
-      })
-
-      it('calls markForHandover via licenceService for finalChecks', () => {
-        const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'roUser')
-
-        return request(app)
-          .post('/hdc/send/finalChecks/123')
-          .expect(() => {
-            expect(prisonerService.getOrganisationContactDetails).to.be.calledOnce()
-            expect(licenceService.markForHandover).to.be.calledOnce()
-            expect(licenceService.markForHandover).to.be.calledWith('123', 'roToCa')
-          })
-      })
-
-      it('calls markForHandover via licenceService for approval', () => {
-        const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'caUser')
-
-        return request(app)
-          .post('/hdc/send/approval/123')
-          .expect(() => {
-            expect(prisonerService.getOrganisationContactDetails).to.be.calledOnce()
-            expect(licenceService.markForHandover).to.be.calledOnce()
-            expect(licenceService.markForHandover).to.be.calledWith('123', 'caToDm')
-          })
-      })
-
-      it('calls markForHandover via licenceService for decided', () => {
-        const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'dmUser')
-
-        return request(app)
-          .post('/hdc/send/decided/123')
-          .expect(() => {
-            expect(prisonerService.getOrganisationContactDetails).to.be.calledOnce()
-            expect(licenceService.markForHandover).to.be.calledOnce()
-            expect(licenceService.markForHandover).to.be.calledWith('123', 'dmToCa')
-          })
-      })
-
-      it('does not removeDecision when sending to decided', () => {
-        const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'dmUser')
-
-        return request(app)
-          .post('/hdc/send/decided/123')
-          .expect(() => {
-            expect(licenceService.removeDecision).to.not.be.called()
-          })
-      })
-
-      it('calls markForHandover via licenceService for refusal', () => {
-        const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'caUser')
-
-        return request(app)
-          .post('/hdc/send/refusal/123')
-          .expect(() => {
-            expect(prisonerService.getOrganisationContactDetails).to.be.calledOnce()
-            expect(licenceService.markForHandover).to.be.calledOnce()
-            expect(licenceService.markForHandover).to.be.calledWith('123', 'caToDmRefusal')
-          })
-      })
-
-      it('calls markForHandover via licenceService for return', () => {
-        const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'dmUser')
-
-        return request(app)
-          .post('/hdc/send/return/123')
-          .expect(() => {
-            expect(prisonerService.getOrganisationContactDetails).to.be.calledOnce()
-            expect(licenceService.markForHandover).to.be.calledOnce()
-            expect(licenceService.markForHandover).to.be.calledWith('123', 'dmToCaReturn')
-          })
-      })
-
-      it('calls removeDecision via licenceService for return', () => {
-        const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'dmUser')
-
-        return request(app)
-          .post('/hdc/send/return/123')
-          .expect(() => {
-            expect(licenceService.removeDecision).to.be.calledOnce()
-            expect(licenceService.removeDecision).to.be.calledWith('123', { licence: { key: 'value' } })
-          })
-      })
-
-      it('audits the send event', () => {
-        const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'dmUser')
-
-        return request(app)
-          .post('/hdc/send/return/123')
-          .expect(() => {
-            expect(prisonerService.getOrganisationContactDetails).to.be.calledOnce()
-            expect(auditStub.record).to.be.calledOnce()
-            expect(auditStub.record).to.be.calledWith('SEND', 'DM_USER', {
+            expect(notificationService.send).to.be.calledWith({
               bookingId: '123',
-              transitionType: 'dmToCaReturn',
-              submissionTarget: {
-                deliusId: 'delius',
-                name: 'Something',
-                premise: 'HMP Blah',
-                agencyId: 'LT1',
-              },
+              transition: transitionsForDestinations.addressReview,
+              licence: { licence: { key: 'value' } },
+              prisoner,
+              token: 'token',
+              user: users.caUser,
             })
           })
       })
 
-      it('shows sent confirmation', () => {
+      it('ro user requests final checks', () => {
+        const app = createApp(
+          { prisonerServiceStub: prisonerService, notificationServiceStub: notificationService },
+          'roUser'
+        )
+
+        return request(app)
+          .post('/hdc/send/finalChecks/123')
+          .expect(() => {
+            expect(notificationService.send).to.be.calledWith({
+              bookingId: '123',
+              transition: transitionsForDestinations.finalChecks,
+              licence: { licence: { key: 'value' } },
+              prisoner,
+              token: 'system-token',
+              user: users.roUser,
+            })
+          })
+      })
+
+      it('dm user returns case to ca', () => {
         const app = createApp(
           {
-            licenceServiceStub: licenceService,
             prisonerServiceStub: prisonerService,
             notificationServiceStub: notificationService,
           },
@@ -293,11 +195,19 @@ describe('send', () => {
           .expect(302)
           .expect(res => {
             expect(res.header.location).to.eql('/hdc/sent/CA/dmToCaReturn/123')
+            expect(notificationService.send).to.be.calledWith({
+              bookingId: '123',
+              transition: transitionsForDestinations.return,
+              licence: { licence: { key: 'value' } },
+              prisoner,
+              token: 'token',
+              user: users.dmUser,
+            })
           })
       })
 
       it('should throw if post requested by wrong user', () => {
-        const app = createApp({ licenceServiceStub: licenceService, prisonerServiceStub: prisonerService }, 'caUser')
+        const app = createApp({ prisonerServiceStub: prisonerService }, 'caUser')
 
         return request(app)
           .post('/hdc/send/return/123')
@@ -305,129 +215,19 @@ describe('send', () => {
           .expect(403)
       })
     })
-
-    describe('Notifications', () => {
-      it('Notifies for new RO case', () => {
-        const app = createApp(
-          {
-            licenceServiceStub: licenceService,
-            prisonerServiceStub: prisonerService,
-            userAdminServiceStub: userAdminService,
-            notificationServiceStub: notificationService,
-          },
-          'caUser'
-        )
-
-        return request(app)
-          .post('/hdc/send/addressReview/123')
-          .expect(() => {
-            expect(notificationService.sendNotifications).to.be.calledOnce()
-            expect(notificationService.sendNotifications).to.be.calledWith({
-              prisoner,
-              bookingId: '123',
-              notificationType: 'RO_NEW',
-              sendingUserName: 'ca last',
-              submissionTarget,
-              token: 'token',
-            })
-          })
-      })
-
-      it('Notifies for returned CA case', () => {
-        const app = createApp(
-          {
-            licenceServiceStub: licenceService,
-            prisonerServiceStub: prisonerService,
-            userAdminServiceStub: userAdminService,
-            notificationServiceStub: notificationService,
-          },
-          'roUser'
-        )
-
-        return request(app)
-          .post('/hdc/send/finalChecks/123')
-          .expect(() => {
-            expect(notificationService.sendNotifications).to.be.calledOnce()
-            expect(notificationService.sendNotifications).to.be.calledWith({
-              prisoner,
-              token: 'system-token',
-              notificationType: 'CA_RETURN',
-              submissionTarget,
-              bookingId: '123',
-              sendingUserName: 'ro last',
-            })
-          })
-      })
-
-      it('Notifies for new DM case', () => {
-        const app = createApp(
-          {
-            licenceServiceStub: licenceService,
-            prisonerServiceStub: prisonerService,
-            userAdminServiceStub: userAdminService,
-            notificationServiceStub: notificationService,
-          },
-          'caUser'
-        )
-
-        return request(app)
-          .post('/hdc/send/approval/123')
-          .expect(() => {
-            expect(notificationService.sendNotifications).to.be.calledOnce()
-            expect(notificationService.sendNotifications).to.be.calledWith({
-              prisoner,
-              token: 'token',
-              notificationType: 'DM_NEW',
-              submissionTarget,
-              bookingId: '123',
-              sendingUserName: 'ca last',
-            })
-          })
-      })
-
-      it('Notifies for decided CA case', () => {
-        const app = createApp(
-          {
-            licenceServiceStub: licenceService,
-            prisonerServiceStub: prisonerService,
-            userAdminServiceStub: userAdminService,
-            notificationServiceStub: notificationService,
-          },
-          'dmUser'
-        )
-
-        return request(app)
-          .post('/hdc/send/decided/123')
-          .expect(() => {
-            expect(notificationService.sendNotifications).to.be.calledOnce()
-            expect(notificationService.sendNotifications).to.be.calledWith({
-              prisoner,
-              token: 'token',
-              notificationType: 'CA_DECISION',
-              submissionTarget,
-              bookingId: '123',
-              sendingUserName: 'dm last',
-            })
-          })
-      })
-    })
   })
 })
-
-function createApp({ licenceServiceStub, prisonerServiceStub, notificationServiceStub, userAdminServiceStub }, user) {
+function createApp({ prisonerServiceStub, notificationServiceStub }, user) {
   const prisonerService = prisonerServiceStub || createPrisonerServiceStub()
-  const licenceService = licenceServiceStub || createLicenceServiceStub()
+  const licenceService = createLicenceServiceStub()
   const signInService = createSignInServiceStub()
-  const notificationService = notificationServiceStub || createNotificationServiceStub()
-  const userAdminService = userAdminServiceStub || createUserAdminServiceStub()
 
   const baseRouter = standardRouter({ licenceService, prisonerService, audit: auditStub, signInService })
   const route = baseRouter(
     createRoute({
       licenceService,
       prisonerService,
-      notificationService,
-      userAdminService,
+      notificationService: notificationServiceStub,
       audit: auditStub,
     }),
     'USER_MANAGEMENT'
